dev_requirements.txt
coverage
djangocms_helper
flake8
ipdb
isort
